Move more sending things to session
Setting spawnpoint done by /checkpoint, fallback for bufferedblocksender, updating entity positions
